cockpit/requirements.txt
absl-py==2.3.1
astunparse==1.6.3
audioread==3.1.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
decorator==5.2.1
flatbuffers==25.9.23
gast==0.4.0
google-auth==2.43.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.76.0
h5py==3.15.1
idna==3.11
jax==0.4.30
jaxlib==0.4.30
joblib==1.5.2
keras==2.12.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.1
Markdown==3.10
MarkupSafe==3.0.3
ml_dtypes==0.5.4
msgpack==1.1.2
numba==0.62.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.2.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.4
wrapt==1.14.2
